backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# NLP and ML
scikit-learn==1.3.2
transformers==4.35.2
torch==2.2.0
nltk==3.8.1

# File parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# PDF generation
reportlab==4.0.7

# CORS (included in starlette/fastapi)
